fun-apps/live-wallpaper/requirements.txt
pygame>=2.5.0
pillow>=12.0.0
